requirements.txt
aniso8601==9.0.1
backports.entry-points-selectable==1.2.0
click==8.1.6
distlib==0.3.7
django-environ==0.10.0
filelock==3.12.2
Flask==2.0.2
Flask-Cors==3.0.10
Flask-RESTful==0.3.9
greenlet==1.1.2
itsdangerous==2.0.1
Jinja2==3.0.3
MarkupSafe
platformdirs==2.2.0
prompt-toolkit==3.0.19
pyarrow==7.0.0
python-dateutil==2.8.2
pytz==2021.1
six==1.16.0
vine==5.0.0
virtualenv==20.7.2
wcwidth==0.2.6
Werkzeug==2.0.2
